Initial development on networking
In order to do what little networking there is, first you need to build
and run (make sure you change the settings so it runs when inactive).

You will be running both the build AND the unity test. On either the
test or the build, create game, input some text, and then create game.
On the other game instance, go to join game and click the game name.
That's all I have so far